# examples/02-ecommerce-service/api/rbac-service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "engine-access-control-service",
    "version": "0.1.0"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/users": {
      "get": {
        "summary": "列出用户列表 [SINCE v0.1.0].\n[EN] List users [SINCE v0.1.0].",
        "operationId": "AccessControlService_UserList",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiUserListResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "page_request.page",
            "description": "可选, 取值:\u003e=0, 默认值为1.\n[EN] Optional, start page, value: \u003e = 0, the default value is 1.",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "page_request.size",
            "description": "数量, 默认取值范围[1,100], 如果超出范围, 则返回失败.\n[EN] Size, default value range [1,100], if it is out of the range, error\nwill be returned.",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "page_request.total",
            "description": "可选, 请求无须填此参数, 响应时系统填写.\n[EN] Optional, this parameter is not required for request, but will be\nfilled in response.",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "sort.field",
            "description": "排序字段, 接口对应字段, 由对应接口解析.\n[EN] Sort field, the corresponding field of the interface, which is parsed\nby the corresponding interface.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "sort.reversed",
            "description": "可选, 是否倒序\n[EN] Optional: Reverse order or not",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "query.email",
            "description": "用户邮箱地址.\n[EN] User email address.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "query.phone",
            "description": "用户手机号码.\n[EN] User phone number.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "query.status",
            "description": "用户状态.\n[EN] User status.\n\n - USER_STAT_UNKNOWN: 未知.\n[EN] Unknown.\n - USER_STAT_NORMAL: 正常.\n[EN] Normal.\n - USER_STAT_DISABLE: 禁用.\n[EN] Disable.",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": [
              "USER_STAT_UNKNOWN",
              "USER_STAT_NORMAL",
              "USER_STAT_DISABLE"
            ],
            "default": "USER_STAT_UNKNOWN"
          },
          {
            "name": "query.periods.start",
            "description": "开始时间, 区间包含.\n[EN] Start time, interval inclusive.",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "query.periods.end",
            "description": "结束时间, 区间不包含.\n[EN] End time, interval exclusive.",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date-time"
          }
        ],
        "tags": [
          "AccessControlService"
        ]
      },
      "post": {
        "summary": "创建用户 [SINCE v0.1.0].\n[EN] Create a user [SINCE v0.1.0].",
        "operationId": "AccessControlService_UserNew",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiUserNewResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "创建用户请求 [SINCE v0.1.0].\n[EN] Create user request [SINCE v0.1.0].",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiUserNewRequest"
            }
          }
        ],
        "tags": [
          "AccessControlService"
        ]
      }
    },
    "/v1/users/{user_id}": {
      "get": {
        "summary": "获取指定用户 [SINCE v0.1.0].\n[EN] Get the specified user information [SINCE v0.1.0].",
        "operationId": "AccessControlService_UserGet",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiUserGetResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "user_id",
            "description": "用户唯一ID,用于保证幂等操作.\n[EN] User unique ID, used to ensure idempotent operation.",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "AccessControlService"
        ]
      },
      "delete": {
        "summary": "删除指定用户 [SINCE v0.1.0].\n[EN] Delete user [SINCE v0.1.0].",
        "operationId": "AccessControlService_UserDelete",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiUserDeleteResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "user_id",
            "description": "用户唯一ID,用于保证幂等操作.\n[EN] User unique ID, used to ensure idempotent operation.",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "AccessControlService"
        ]
      },
      "patch": {
        "summary": "更新指定用户信息 [SINCE v0.1.0].\n[EN] Update information of the specified user [SINCE v0.1.0].",
        "operationId": "AccessControlService_UserUpdate",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiUserUpdateResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "user_id",
            "description": "用户唯一ID,用于保证幂等操作.\n[EN] User unique ID, used to ensure idempotent operation.",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AccessControlServiceUserUpdateBody"
            }
          }
        ],
        "tags": [
          "AccessControlService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "AccessControlServiceUserUpdateBody": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "description": "用户邮箱地址.\n[EN] User email address."
        },
        "phone": {
          "type": "string",
          "description": "用户手机号码.\n[EN] User phone number."
        },
        "avatar": {
          "type": "string",
          "description": "用户头像.\n[EN] User avatar."
        },
        "nickname": {
          "type": "string",
          "description": "用户昵称.\n[EN] User Nickname."
        },
        "status": {
          "$ref": "#/definitions/apiUserStatus",
          "description": "用户状态.\n[EN] User status."
        },
        "role_ids": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "description": "角色ID列表.\n[EN] Role id list."
        }
      },
      "description": "更新指定用户请求 [SINCE v0.1.0].\n[EN] Update user request [SINCE v0.1.0]."
    },
...
    "access_controlapiUserInfo": {
      "type": "object",
      "properties": {
        "user_id": {
          "type": "string",
          "description": "用户唯一ID,用于保证幂等操作.\n[EN] User unique ID, used to ensure idempotent operation."
        },
        "username": {
          "type": "string",
          "description": "用户名.\n[EN] User name."
        },
        "email": {
          "type": "string",
          "description": "用户邮箱地址.\n[EN] User email address."
        },
        "phone": {
          "type": "string",
          "description": "用户手机号码.\n[EN] User phone number."
        },
        "avatar": {
          "type": "string",
          "description": "用户头像.\n[EN] User avatar."
        },
        "nickname": {
          "type": "string",
          "description": "用户昵称.\n[EN] User Nickname."
        },
        "status": {
          "$ref": "#/definitions/apiUserStatus",
          "description": "用户状态.\n[EN] User status."
        },
        "role_ids": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "description": "角色ID列表.\n[EN] Role ID list."
        },
        "created_at": {
          "type": "string",
          "format": "date-time",
          "description": "创建时间.\n[EN] Created time."
        },
        "updated_at": {
          "type": "string",
          "format": "date-time",
          "description": "上次修改时间.\n[EN] Updated time."
        }
      },
      "description": "用户信息.\n[EN] User information."
    },
...
    "apiUserDeleteResponse": {
      "type": "object",
      "description": "删除指定用户响应 [SINCE v0.1.0].\n[EN] Delete user response [SINCE v0.1.0]."
    },
    "apiUserGetResponse": {
      "type": "object",
      "properties": {
        "user_info": {
          "$ref": "#/definitions/access_controlapiUserInfo",
          "description": "用户信息.\n[EN] User information."
        }
      },
      "description": "用户查询响应 [SINCE v0.1.0].\n[EN] Get user response [SINCE v0.1.0]."
    },
...
    "apiUserListResponse": {
      "type": "object",
      "properties": {
        "page_response": {
          "$ref": "#/definitions/commonapisPaging",
          "description": "分页返回信息\n[EN] Page response."
        },
        "user_infos": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/access_controlapiUserInfo"
          },
          "description": "用户信息列表.\n[EN] User informations."
        }
      },
      "description": "用户列表响应 [SINCE v0.1.0].\n[EN] User listing response [SINCE v0.1.0]."
    },
...
    "apiUserNewRequest": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string",
          "description": "用户名.\n[EN] User name."
        },
        "email": {
          "type": "string",
          "description": "用户邮箱地址.\n[EN] User email address."
        },
        "phone": {
          "type": "string",
          "description": "用户手机号码.\n[EN] User phone number."
        },
        "password": {
          "type": "string",
          "format": "byte",
          "description": "用户密码,用于登录系统.\n[EN] User password, used to log in to the system."
        },
        "avatar": {
          "type": "string",
          "description": "用户头像.\n[EN] User avatar."
        },
        "nickname": {
          "type": "string",
          "description": "用户昵称.\n[EN] User Nickname."
        },
        "status": {
          "$ref": "#/definitions/apiUserStatus",
          "description": "用户状态.\n[EN] User status."
        },
        "role_ids": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "description": "角色ID列表.\n[EN] Role id list."
        }
      },
      "description": "创建用户请求 [SINCE v0.1.0].\n[EN] Create user request [SINCE v0.1.0]."
    },
    "apiUserNewResponse": {
      "type": "object",
      "properties": {
        "user_id": {
          "type": "string",
          "description": "用户唯一ID,用于保证幂等操作.\n[EN] User unique ID, used to ensure idempotent operation."
        }
      },
      "description": "创建用户响应 [SINCE v0.1.0].\n[EN] Create user response [SINCE v0.1.0]."
    },
...
    "apiUserStatus": {
      "type": "string",
      "enum": [
        "USER_STAT_UNKNOWN",
        "USER_STAT_NORMAL",
        "USER_STAT_DISABLE"
      ],
      "default": "USER_STAT_UNKNOWN",
      "description": "用户状态.\n[EN] User status.\n\n - USER_STAT_UNKNOWN: 未知.\n[EN] Unknown.\n - USER_STAT_NORMAL: 正常.\n[EN] Normal.\n - USER_STAT_DISABLE: 禁用.\n[EN] Disable."
    },
    "apiUserUpdateResponse": {
      "type": "object",
      "description": "更新指定用户响应 [SINCE v0.1.0].\n[EN] Update user response [SINCE v0.1.0]."
    },
...
    "commonapisPaging": {
      "type": "object",
      "properties": {
        "page": {
          "type": "integer",
          "format": "int32",
          "description": "可选, 取值:\u003e=0, 默认值为1.\n[EN] Optional, start page, value: \u003e = 0, the default value is 1."
        },
        "size": {
          "type": "integer",
          "format": "int32",
          "description": "数量, 默认取值范围[1,100], 如果超出范围, 则返回失败.\n[EN] Size, default value range [1,100], if it is out of the range, error\nwill be returned."
        },
        "total": {
          "type": "integer",
          "format": "int32",
          "description": "可选, 请求无须填此参数, 响应时系统填写.\n[EN] Optional, this parameter is not required for request, but will be\nfilled in response."
        }
      },
      "description": "通用翻页结构.\n[EN] Common paging structure."
    },
...
    "googlerpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
...
  }
}